Added Commnet Deletion, Added Sample Data

+++ Areas/PostArea/Controllers/CommentController.cs
@@ -25,12 +25,25 @@ namespace SocialMediaProject.Areas.PostArea.Controllers
             return View(model);
         }
 
+        [HttpPost]
         public IActionResult Delete(int id)
         {
-            var comment = _dbContext.Comments.Find(id);
-            _dbContext.Comments.Remove(comment);
-            _dbContext.SaveChanges();
-            return RedirectToAction("View", "Post", new { id = comment.PostId, area = "PostArea" });
+            var comment = _dbContext.Comments.FirstOrDefault(c => c.Id == id);
+
+            if (comment == null)
+            {
+                return NotFound();
+            }
+
+            if (TempData["UserId"] is int userId && comment.UserId == userId)
+            {
+                _dbContext.Comments.Remove(comment);
+                _dbContext.SaveChanges();
+
+                return RedirectToAction("View", "Post", new { id = comment.PostId, area = "PostArea" });
+            }
+
+            return Forbid();
         }
     }
 }
